dataName 달아주기까지

+++ src/main/java/com/kos/exam/boot/controller/UsrArticleController.java
@@ -32,13 +32,13 @@ public class UsrArticleController {
 			loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
 		}
 		if(isLogined == false) {
-			return ResultData.from("F-A", "로그인 후 이용해주세요", null);
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
 		if(Ut.empty(title)) {
-			return ResultData.from("F-1", "title을 입력해주세요", null);
+			return ResultData.from("F-1", "title을 입력해주세요");
 		}
 		if(Ut.empty(body)) {
-			return ResultData.from("F-1", "body을 입력해주세요", null);
+			return ResultData.from("F-1", "body을 입력해주세요");
 		}
 		
 		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
@@ -46,7 +46,7 @@ public class UsrArticleController {
 		
 		Article article = articleService.getArticle(id);
 		
-		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), article);
+		return ResultData.from(writeArticleRd.getResultCode(), writeArticleRd.getMsg(), "Article", article);
 	}
 	
 	@RequestMapping("/usr/article/getArticles")
@@ -54,7 +54,7 @@ public class UsrArticleController {
 	public ResultData<List<Article>> getArticles() {
 		List<Article> articles =  articleService.getArticles();
 		
-		return ResultData.from("S-1", "게시물 리스트입니다.", articles);
+		return ResultData.from("S-1", "게시물 리스트입니다.","List Articles", articles);
 	}
 	
 	@RequestMapping("/usr/article/getArticle")
@@ -63,10 +63,10 @@ public class UsrArticleController {
 		Article article = articleService.getArticle(id);
 		
 		if ( article == null ) {
-			return ResultData.from("F-1", Ut.f("%d게시물이 존재하지 않습니다", id),null);
+			return ResultData.from("F-1", Ut.f("%d게시물이 존재하지 않습니다", id));
 		}
 		
-		return ResultData.from("S-1", Ut.f("%d번 게시물입니다.", id),article);
+		return ResultData.from("S-1", Ut.f("%d번 게시물입니다.", id),"Article",article);
 	}
 	
 	@RequestMapping("/usr/article/doDelete")
@@ -80,34 +80,53 @@ public class UsrArticleController {
 			loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
 		}
 		if(isLogined == false) {
-			return ResultData.from("F-A", "로그인 후 이용해주세요", null);
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
 		}
 		
 		Article article = articleService.getArticle(id);
 		
 		if ( article == null ) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다", id), null);
+			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
+		}
+		
+		if(article.getMemberId()!= loginedMemberId) {
+			return ResultData.from("F-2", "권한이 없습니다.");
 		}
 		
 		articleService.deleteArticle(id);
 		
-		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제하였습니다.", id), id);
+		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제하였습니다.", id),"회원번호", id);
 	}
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int)httpSession.getAttribute("loginedMemberId");
+		}
+		if(isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+		
 		Article article = articleService.getArticle(id);
 		
 		if ( article == null ) {
-			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다", id), null);
+			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 		}
 		
-		articleService.modifyArticle(id, title, body);
+		ResultData actorCanModifyRd = articleService.actorCanModify(loginedMemberId, article);
 		
-		return ResultData.from("F-1", Ut.f("%d번 게시물을 수정하였습니다", id), null);
+		if(actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+		
+		return articleService.modifyArticle(id, title, body);
 	}
 	// 액션 메서드 끝
-	
-	
+
 }
